Lots more fun player movement changes.
I will get around to making some of the player movement shorter and more
readable, and probably in different file/s.

One of the main / most interesting things I added was a small amount of
air control, so that when you are in the air you can affect the
movement, but only slightly. Play around with the playerAirControl
variable in main.cpp if you're interested.

We need a way to generate levels? That list of cubes is getting rather
long!

Replaced some for(i, i < x i++) loops with for(auto i : x) loops,
because they are nice.